Almost 1.4.8 - Jabberwock Complete, Spawning Fix
The Jabberwock is pretty much good to go along with the Weight effect.
The mob spawning issues seem to be resolved finally! I have removed the
Chunk Spawn Limit config option and the method is now passed to vanilla
returning 4 by default, I'm not sure what it actually do but it doesn't
seem to be a Chunk Spawn Limit!
Some spawn defaults have been tweaked and feel pretty nice now.
The Mountain Mobs sub-mod is good to go too!
This is just about a release, but some more tests could be done and
possibly some improvements here and there.
TODO Ventroraptors and Uvaraptors could also be meat tamed specific.