Update 1.4.9 for MC1.6.4 BETA
This is basically ready to be released, but I'm tired and it's late so I
should double check things in the morning before releasing! :P
Fixed Jousts and other mobs sometimes not dropping things.
Frostweaver can no longer fly and Frostwebs can now be placed correctly.
Created the textures for the Frostweb items and block.
The Erepede's sand speed boost now only works when they are mounted,
they were far to overpowered as regular mobs.
Adding climbing to the Concapede and web resistance and climbing to the
Frostweaver.