Implemented MobInfo and SpawnInfo Classes
Further work still needs to be done as well as vigorous testing!
I've also made use of the forSpawnCount boolean on the
EntityCreatureBase.isCreatureType() method, this could be the source of
lack of mob spawns!
I still also need to work on mob persistence as farmable mobs seem to
still be despawning.