Troll is in game, 3D projectiles next!
The Troll is in game with it's stone skin ability in daylight, sounds,
model and animation all working nicely, they fire Venom Shot projectiles
for now as a placeholder.
The next big job are the 3D projectiles, once they're done some fix ups
and 1.4.10 is good to go!
I also need to double check some mobs as they seem to show up at day a
lot.