Better Pet Control
Pet controls are working very nicely now, but there are still some
imperfections.
Sitting On and Passive Off should turn pets into guards where they will
guard the spot they are sitting at, chase and attack enemies and then
return to their post after or if the enemy gets too far.
Aggressive stance should enable some new hunting AI that seeks out any
mob that would pose a threat to the player.
The ultimate goal is to use Sitting On, Passive Off, Stance Aggressive
and PVP On to have  pet guard a players house and attack other players
or mobs on sight should they get too close, even when the owner is not
on the server.

The summoning portal and laser beam end still have a weird block
raytracing issue that needs to be solved. Could it be the starting Y
height?