Update 1.10.2.4: It is pitch black.
Improvement: All Cooked Food items such as Cooked Aspid Meat now provide
a buff again, the duration is not as long as the advanced foods though
such as Moss Pie.

Bug Fix: The Shadow Games event now jumps the world time to the next
night rather than resetting the time to the first ever night.
Bug Fix: Mobs that spawn in the sky through the custom spawner should
now respect the minimum range from the player on the Y axis.

I've also added in a couple of new NBTTags for other devs or map editors
to play with, they may not be 100% functional yet though. See the
api.info.ExtraMobBehaviour class on GitHub for all the NBTTags that can
be used in the ExtraBehaviour compound tag.