Update 1.10.4.5
Bug Fix: Mobs added to custom spawners should now all spawn (only the
first and last entries were being randomly chosen hence some events
don't have all of their mobs and some mobs such as the Skylus or Ika
weren't spawning).

Improvement: Lowered the Demon Mob spawn weights from the likes of 100
to 8 to match other mobs, however if any mob is added to the Nether
(Hell) Biome, their weight will be multiplied by 10, this is because
vanilla Nether mobs such as the Zombie Pigmen have a weight of 100, I'm
not sure why this is but this is a cool work around to keep the Demon
Mobs balanced when used in Events and when set to spawn.