Tweaks
Bug Fix: Tweaked the darkness spawn type some more as it still seems to
be problematic. It will now also print darkness level updates to the
console if the CustomSpawner debug option is set to true.

Bug Fix: Some mobs were not in their groups such as Frostweavers not
being in the Ice group and Khlaks not being in the fire group (where
fire and ice mobs fight each other).