Tweaks
Major Bug Fix: Completely overhauled the event timing system. Previously
is was about 6x faster than it should have been and was also very
limited and inefficient. It is now uses the total world time (or current
if set in the config) which makes it accurate. Because of this the
config defaults have been changed and thus the config will reset.

Improvement: The Verdant Cinderfall Sword and Verdant Venom Axeblade now
have a higher chance of summoning minions instead of an extended minion
duration.

Improvement: Cinders and Remobra summoned from the minion swords are now
scaled own to 40% their normal size to keep them out of the way.

Improvement: The Ember projectile (cast by Cinders and the Ember
Scepter) has been drastically reduced in size to improve visibility.

Improvement: The Venom Axeblades will now grant players immunity to
poison if they are holding them in their hand, this is very handy as
they summon Remobra that create poison clouds.

Improvement:
The Item Lists Config now has semicolon separated entries using commas
for each part of the entries in order to be consistent with every other
config input that uses semicolons and commas.

Improvement: The texture color of Geist Liver has been changed to match
the color scheme of Shadowfire.