1.11.0.3 Progress
Added rare health bar config.
Fixed blizzard charges in dispensers.
Predators such as Chupacabras or Ventoraptors will no longer attack mobs
that are both classed as both an Animal/Alpha and Predator. This will
stop unintended targetting such as a Ventoraptor attacking a Pinky or a
Shade attacking a Chupacabra.