1.12.0.0 Progress
Created an automatic Hit Area system that creates multiple hit boxes for
large entities ranging from Trents and Striders up to Rahovart, the hit
areas can also be larger than the main collision box, this is used for
Rahovart as well as Crusks, Serpix and Astaroth and does not effect
other collision such as block collision.

The Hellfire Charges and Wall attack still need fixing.
The Hellfire Barrier needs to be implemented.

Not much left to do now overall though!