1.12.0.0 Progress
Hellfirewall Working! Next up are the Hellfire Barriers which should be
fairly simple now.
The rotations are still bugged for the Hellfire Charges and the Wall
Rotation so that will need to be addressed too.
The 'Eerie' Cacodemon will need to be added as will the Demonic
Soulstone, I may also look into creating some custom Demonic Brick
Blocks, Tiles and Stone for the arena, I should look into having
mini-arenas naturally generate in the nether with a random demon mob
spawner in the middle too!
Also made Remobras always drop Poison Glands.