1.14.0.0 Demon Mobs
Demon Mobs fully implemented.
Fixed errors with Astaroth and Trite model material files.
Fixed a case-sensitive issue with the Ika model material file, this
might be the cause of it being invisible for some and with Optifine.
Fixed an error where some fire blocks would not age due to their age
speed being an odd number (integers don't like to divide into decimals
and need to be rounded from floats or doubles).
Fixed a bug where Altars were not working.