1.14.2.0 Jumping Asmodeus and New Strider
Asmodeus now leaps around the arena in pursuit of it's current primary
target, there is no escape!

The Strider has been redesigned, it now has a new model, it is half it's
old size (was too big for how common it was) and it now acts as a normal
pet again instead of a mount.