1.14.3.2 Improvements
Poison Clouds will not fade away if they are on top of Mycelium.

Fixed a bug where the block spawner surface was only working for
material entries.

Added a second random check for the Lunar spawner in the End to make
spawning more ellusive.

Shadowstone and Demonstone now costs 8 Cobblestone and 1 Nether
Warts/Obsidian and creates 9 blocks.