THIS commit fixes the glitchy rotation!
The size of the entity affects the position tracking on the server
(which is only synced to the nearest 1/32 of a block).  The server
warps the entity on the client to the wrong position if the size does
not match.  The client sees this as a rapid movement and so resets the
rotation of the entity.