AI Follow Master seemed broken
With the debug text open I found groups of stationery Makas were stuck
on AIFollowMaster.  The pathfinding was failing to locate the master
because they are outside of the path finder’s range.  I found that
range was limited to the value of SharedMonsterAttributes.followRange.
We have a few options.
1) increase followRange to equal the max distance you want Makas to
find their Alphas
2) decrease the distance on the AIFollowMaster
3) Write a custom PathFinder
I think what I have done here is the most simple but a PathFinder with
a large range might be slow (have not done any load testing).
This will probably need checking for all mobs that follow other
entities.