New Config Features
Added global config settings to control if Mob Events can affect weather or the time of day.

Added a minimum world day spawn condition to each mob. For example you can now configure it so that Lobbers cannot spawn until the world is at least 10 days old.

Added a config entry to allow for hitbox scaling of each mob, this is relative to the hard coded internal size.

The Roa Whirlpool setting is now the range (in blocks) of the wirlpool rather than true/false, setting it to 0 will disable it completely.

The Roa's whirlpool can no longer affect any entities with the Weight effect active on them.