1.16.2.2 Massive Bugfixes, Improvements, New Ent
The Ent's model and texture has been improved!
The Ventoraptors and Uvaraptors now have deadly sickle claws!

Added a new FISHING spawn type! Yup, now when you real in a catch, you may get a lot more than you bragained for! Configs will need to be manually cleared or updated with the FISHING type for this to work. The default mobs that spawn from this are: Silex, Abtu, Lacedon and Roa.
Added the Shadow Clown skin to the Grue, these will spawn on halloween and Grues with the name "Shadow Clown" will use this skin.

Fixed seasonal mobs skins not being applied.