1.17.0.1 Pathing Overhaul
============================================================
Lycanites Mobs Update: Ultimus - Version 1.17.0.1 for Minecraft 1.11.2
Configs older than 1.14.2.2 will be reset.
============================================================

--------------------
Major Fixes:
--------------------
I've rewrote the entire path navigation, path node processor and move helper system. This should fix a lot of pathing issues with mobs and it also means that they all consistently share the same code which adapts based on the mob, making it much easier to maintain as easier to deal with hybrids such as the Lacedon.
I've tried to be very performance conscious with this also. This is very new and will need testing but looks very stable! The most noticeable fixed mobs are the Lacedon, Dweller and Herma. The lava mobs such as the Lobber seem to move around in lava better too but this needs more testing.
Fixed a bug where flying and swimming mobs weren't updating their limb movements, this isn't noticeable on most of these mobs as the limbs use different animations but the Skylus, Lacedon, etc would begin to twitch when hit because of this. 


--------------------
Changes:
--------------------
All mounts will no longer dismount the player when in water, most of them will however ignore player commands and swim back to shore themselves.