1.17.0.3 Spawn and Pickup Fixes
============================================================
Lycanites Mobs Update: Ultimus - Version 1.17.0.3 for Minecraft 1.11.2
Configs older than 1.14.2.2 will be reset.
============================================================

--------------------
Major Fixes:
--------------------
Fixed a crash caused by events that have no mobs (should they be disabled in the config), additionally, the event will no longer try to start if there are no mobs enabled for it to spawn.
Fixed a bug where the bottom dwelling mobs were placed on the water floor for the server but not updating on the client causing them to seemingly float above the water and become immortal!
The pickup logic no longer triggers a teleport on players, this was used to fix a bug with mobs but causes problems with player.
Pickup offsets will now check if the location it is offsetting the picked up entity to is a solid block, if it is then the offset is changed to 0, this should reduce the odds of players getting dragged into blocks and glitching about.


--------------------
Minor Fixes:
--------------------
Fixed Strider recipe (was accidentally removed)!
Raiko should apply the weight debuff when randomly dropping their prey again.
Skylus have had their shell-less speed lowered and will no longer has the weight applying code that the Raiko had (was bugged and never worked anyway).