JSON Spawner Logic Complete
All logic for the JSON Spawners is in place, I now just need to load them up from JSON files as well as recreate all the custom spawners in JSON format.

This is all untested so there may be problems but the code is pretty solid so hopefully there will be no major problems.