Equipment Forge Start
Started to make the Equipment Forge.
It currently has an inventory and opens a GUI.
I need to make a proper GUI for it.
I need to have the Equipment Part slots work with it correctly.
I need to get the mode/confirmations working in the GUI.

After that I need to create Equipment as well as a model renderer for Equipment.