Dungeon Fixes and Soulstone Fixes
Further fixed up the new dungeon system, changed rotation to facing using EnumFacing for better usability, collisionSize is renamed to occupiedSize for better understanding, wrapped roomSize and occupiedSize into getters and fixed the main dungeon entrance.
Stairs are now working but are still not themed.
Loaded (from world NBT) dungeon generation now works.
WorldGen is still cascading a little bit, there is no noticeable lag but this is not idea for possible custom oversized dungeons.
Entrnaces on chunk borders aren't clearing fully, this is likely becasue the getChunkPositions() method is wrong, this needs to be tested with a gold border or something and might be the cascading worldgen cause.
Boss sectors and boss mob entries need added.
Inner Stairs Wall layers are cutting into the corners a bit (or is this for any small room)?