Village Spawn Location and Village Protection AI
Added new Village Spawn Type for spawning mobs in Villages.
Added new Defend Village AI based on the Iron Golem.
The Aegis will now monitor the nearest attackable player and will fixate if that player breaks chests, furnaces or doors as well as if the player opens a chest or furnace container!