Complete Blending Fix
Finally fixed all the texture blending issues!
Reduced the projectile speed of Vapula, Afrit and Arix attacks.
Doubled the time it takes for a Vapula to recharge.
Trying to fix this weird Pet Owner UUID bug, it works great in the development environment but when built, setting or getting an owner via player UUID becomes inconsistent.