Player Owner Bug
I've tried all sorts but still can't track down this player owner bug, for unknown reasons the player owner entity/uuid is returning as null when called but yet still shows as not null at the same time.
This only happens with compiled versions of the mod and is not related to online or offline players either which makes it a huge pain to test as I have to compile the entire mod and run it every time I want to test a change.