Equipment Models Working!
Can now successfully connect parts together to create a full Equipment Piece.
Parts need to be consumed when the piece is taken.
The client piece slot needs updated.
Existing pieces need to be editable.
Slots need to lock when editing.
Features need to be implemented.
Naming needs to be added.
Tooltips need to be looked at, perhaps use keyboard modifiers to reveal types of info.